+++ src/async_impl/mod.rs
@@ -2,6 +2,7 @@ pub use self::body::Body;
 pub use self::client::{Client, ClientBuilder};
 pub use self::request::{Request, RequestBuilder};
 pub use self::response::Response;
+pub use self::upgrade::Upgraded;
 
 #[cfg(feature = "blocking")]
 pub(crate) use self::decoder::Decoder;
@@ -13,3 +14,4 @@ pub mod decoder;
 pub mod multipart;
 pub(crate) mod request;
 mod response;
+mod upgrade;